+++ opaca-model/src/main/java/de/gtarc/opaca/model/Action.java
@@ -1,9 +1,10 @@
 package de.gtarc.opaca.model;
 
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 
 import java.util.Map;
 
@@ -14,17 +15,18 @@ import java.util.Map;
 public class Action {
 
     /** name of the action */
-    @NonNull
+    @NotNull
     String name;
 
     /** optional human-readable description of what this action does */
     String description;
 
     /** parameter names and types */
-    @NonNull
+    @NotNull @Valid
     Map<String, Parameter> parameters = Map.of();
 
     /** type of result */
+    @Valid
     Parameter result;
 
     public Action(String name, Map<String, Parameter> parameters, Parameter result) {