+++ opaca-model/src/main/java/de/gtarc/opaca/model/Parameter.java
@@ -2,10 +2,11 @@ package de.gtarc.opaca.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 
 /**
  * Describes parameters of an Action
@@ -14,34 +15,41 @@ import lombok.NonNull;
 public class Parameter {
 
     /** the type name; either a primitive, or 'array', or defined in definitions or definitionsByUrl of image */
-    @NonNull
+    @NotNull
     String type;
 
     /** whether the parameter is required; if it's not, the default may be determined by the action itself */
-    Boolean required = true;
+    boolean required = true;
+
+    /** default value for the parameter; ignored if required, otherwise optional, should be of given type */
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    Object defaultValue = null;
 
     /** if type is 'array', this is the type of the array's items */
     @JsonInclude(JsonInclude.Include.NON_NULL)
+    @Valid
     ArrayItems items = null;
 
+    /** short-hand for a required parameter */
     public Parameter(String type) {
         this.type = type;
     }
 
-    public Parameter(String type, Boolean required) {
+    /** short-hand for an optional parameter, with or without default */
+    public Parameter(String type, Object defaultValue) {
         this.type = type;
-        this.required = required;
+        this.required = false;
+        this.defaultValue = defaultValue;
     }
 
     @Data @AllArgsConstructor @NoArgsConstructor
     public static class ArrayItems {
-        @NonNull
+        @NotNull
         String type;
 
         @JsonInclude(JsonInclude.Include.NON_NULL)
+        @Valid
         ArrayItems items = null;
     }
 
 }
-
-